requirements.txt
django==2.2.7
django-cors-headers==3.2.0
djangorestframework==3.10.3
pillow==6.2.1
mysqlclient
mysql-python